feat: better structure of log io
