+++ internal/cmd/cli/streams/agent-logs.json
@@ -1,6 +1,10 @@
 {
     "name": "agent-logs",
-    "subjects": ["NITS.AGENT.*.LOG", "NITS.AGENT.*.LOG.>", "NITS.AGENT.*.OUT.>"],
+    "subjects": [
+        "NITS.AGENT.*.LOG",
+        "NITS.AGENT.*.LOG.>",
+        "NITS.AGENT.*.OUT.>"
+    ],
     "retention": "limits",
     "max_consumers": -1,
     "max_msgs_per_subject": -1,
